requirements.txt
Pillow==2.8.1
clarifai==0.2
flickrapi==2.0
oauthlib==0.7.2
requests==2.7.0
requests-oauthlib==0.5.0
requests-toolbelt==0.4.0
six==1.9.0
wsgiref==0.1.2
